refactor: separate dependencies into a single file
Functions for dependency injections and helper functions were all over
the place. Now, they are all defined in dependencies.py. Import were changed
to match this.
